This commit failes on N-Queens 4
A serious bug has been found dealing with the stack Chunks.

When the GCode instruction 'Pop' is executed, it is possible
that the new stack pointer will point to a different chunk than
the Machine's stackObj pointer. There are a few possible fixes,
the easiest will be that instead of using 'getNthAddrFrom' just
use a loop that uses 'stackPopThrowAway' N times.

A better mether might be to create a new function that looks like
'getNthAddrFrom' but takes into account the cases where the Nth
address will be the new stack pointer.