Made par switch array work more sensibly
When being passed to a function (parseGCode) the par switch array suffered
from the classic 'value being modified in the function's scope does not
change the value in the caller's scope'.

To remedy this I 'passed by reference' so that I could get the new value
without having to return it from the function.

I feel dirty.