Abstracted out the execution of a program.
We now have an 'executeProg' function that can be called by our searching
functions. This means that we can now treat the execution of a program more
like a black box (as it should have been done in the beginning).

Right now there are only two options available from the CLI and they don't
do much.

'-I <int>': Maximum number of iterations for the search. If another option that
     controls this is passed after 'I' on the CL, then 'I' is overridden.

'-R <int>': Use random search and iterate the passed number of times.
     Right now, the random search doesn't do anything! That needs to change.