Fixed GC bug that copied CONSTR fields incorrectly
Basically, we were trying to copy from the *to* space instead of from the
*from* space.

We've also added a Heap Overflow error in case your data structures are too
big. Eventually we will add a flag that lets you specify how large the heap
should be and whether you allow it to grow.