Now marking some sub-expressions in WHNF.
This allows us to return partial functions without the application
nodes being locked. Some questions arise:

Can we optimise Eval so that we don't try to reduce expressions in
WHNF?

Can we know that there is never a reason to _unlock_ a node and
restore it to a normal APP (i.e. does it always become a WHNF_APP?)

I think the answer to the first one is yes and the second one is
probably...